% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{adding}
\alias{adding}
\title{An adding function}
\usage{
adding(x, y)
}
\arguments{
\item{x}{A numeric value, scalar or vector}

\item{y}{A numeric value, scalar or vector}
}
\value{
The sum of the two arguments
}
\description{
An adding function
}
\examples{
adding(1, 1)
adding(c(1,1),c(2, 3))


}
