% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello.R
\name{hello}
\alias{hello}
\title{Hello,world}
\usage{
hello(times = 1)
}
\arguments{
\item{times}{The number of times to print}
}
\value{
a ggplot object
}
\description{
This is an example function named Hello.R, which prints `Hello, World!!`
}
\examples{
temp1 <- hello(times=1);
temp2 <- hello();
}
