% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myloop.R
\name{myloop}
\alias{myloop}
\title{Print days of the week}
\usage{
myloop(i)
}
\arguments{
\item{i}{Charactor vector of some days of the week}
}
\value{
A charactors vector
}
\description{
Print days of the week
}
\examples{
myloop(c("Mon", "Tue", "Wed"))
}
